slot edits component; add passback update
