+++ src/components/dashboard/containers/Slot/SlotBasic.js
@@ -1,11 +1,8 @@
 import React, { Fragment } from 'react'
-import { useSelector } from 'react-redux'
 import { Paper, Grid, Box } from '@material-ui/core'
 import { WebsiteIssues } from 'components/dashboard/containers/Slot/WebsiteIssues'
 import TargetsList from 'components/dashboard/containers/TargetsList'
-import WithDialog from 'components/common/dialog/WithDialog'
-import FormSteps from 'components/common/stepper/FormSteps'
-import { AdSlotTargeting } from 'components/dashboard/forms/items/NewItems'
+import { SlotEdits } from './SlotEdits'
 import {
 	DirtyProps,
 	ItemTitle,
@@ -16,57 +13,7 @@ import {
 	ItemMinPerImpression,
 	ItemWebsite,
 } from 'components/dashboard/containers/ItemCommon/'
-import { t, selectNewItemByTypeAndId } from 'selectors'
-import {
-	execute,
-	updateNewItem,
-	resetNewItem,
-	updateSlotTargeting,
-} from 'actions'
-import { AdSlot } from 'adex-models'
-
-// TODO: Targeting will be changed, but this will be used for campaigns targeting update
-export const TargetingSteps = ({ updateField, itemId, ...props }) => {
-	return (
-		<FormSteps
-			{...props}
-			cancelFunction={() => execute(resetNewItem('AdSlot', itemId))}
-			itemId={itemId}
-			validateIdBase='new-AdUnit-'
-			itemType='AdSlot'
-			stepsId='new-slot-'
-			steps={[
-				{
-					title: 'SLOT_TAGS_STEP',
-					component: AdSlotTargeting,
-					completeBtnTitle: 'OK',
-					completeFn: props =>
-						execute(updateSlotTargeting({ updateField, itemId, ...props })),
-				},
-			]}
-			itemModel={AdSlot}
-		/>
-	)
-}
-
-const TargetEdit = WithDialog(TargetingSteps)
-
-const getInitialTemp = ({ temp, tags = [] }, currentTags, dirtyProps = []) => {
-	const targets = dirtyProps.includes('tags')
-		? currentTags
-		: [...tags].map((tag, key) => ({
-				key,
-				collection: 'tags',
-				source: 'custom',
-				label: t(`TARGET_LABEL_TAGS`),
-				placeholder: t(`TARGET_LABEL_TAGS`),
-				target: { ...tag },
-		  }))
-
-	const newTemp = { ...temp, targets }
-
-	return { temp: newTemp }
-}
+import { t } from 'selectors'
 
 export const SlotBasic = ({ item, ...hookProps }) => {
 	const {
@@ -79,21 +26,6 @@ export const SlotBasic = ({ item, ...hookProps }) => {
 		website,
 	} = item
 	const { title: errTitle, description: errDescription } = hookProps.validations
-	const { temp = {} } = useSelector(state =>
-		selectNewItemByTypeAndId(state, 'AdSlot', item.id)
-	)
-
-	const onTargetingClick = () => {
-		execute(
-			updateNewItem(
-				item,
-				getInitialTemp(item, temp.targets, hookProps.dirtyProps),
-				'AdSlot',
-				AdSlot,
-				item.id
-			)
-		)
-	}
 
 	return (
 		<Fragment>
@@ -113,14 +45,9 @@ export const SlotBasic = ({ item, ...hookProps }) => {
 								<ItemFallbackMediaURL targetUrl={targetUrl} />
 							</Box>
 							<Box py={1}>
-								<TargetEdit
-									btnLabel='UPDATE_TAGS'
-									title='UPDATE_SLOT_TAGS'
-									itemId={item.id}
-									disableBackdropClick
-									updateField={hookProps.updateField}
-									onClick={onTargetingClick}
-								/>
+								<SlotEdits item={item} hookProps={hookProps} />
+							</Box>
+							<Box py={1}>
 								<TargetsList targets={item.tags} subHeader={'PROP_TAGS'} />
 							</Box>
 						</Grid>
